pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-modal-reasoning-system"
version = "0.1.0"
description = "A modern multi-modal reasoning system for vision-language tasks"
authors = [{name = "AI Projects", email = "projects@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "tensorboard>=2.13.0",
    "opencv-python>=4.8.0",
    "scikit-image>=0.21.0",
    "Pillow>=10.0.0",
    "transformers>=4.30.0",
    "sentencepiece>=0.1.99",
    "accelerate>=0.20.0",
    "open-clip-torch>=2.20.0",
    "sacrebleu>=2.3.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "nltk>=3.8.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "streamlit>=1.25.0",
    "gradio>=3.40.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.0.280",
    "pytest>=7.4.0",
    "pre-commit>=3.3.0",
]
advanced = [
    "diffusers>=0.20.0",
    "decord>=0.6.0",
    "faiss-cpu>=1.7.4",
    "wandb>=0.15.0",
]

[project.urls]
Homepage = "https://github.com/kryptologyst"
Repository = "https://github.com/kryptologyst/multi-modal-reasoning-system"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
